Add Tokenizer and support classes.
The tokenizer is responsible for the conversion of raw text to a list of
processed tokens which can be used for later stages. It can be customized
with a list of filler words to ignore and a Stemmer to stem words.

The Stemmer interface can be used to extend the behavior of the
tokenizer with custom stemmers.

The StringIntener allows us to do fast string compares at run time.

This patch also adds test for these classes.