Add fact intent implementation
Scan the sentence for the fact key word and tries to locate a subject
token afterwards. This should mostly work since the subject is available
immediately after the keyword.

To support case where the subject is behind the keyword, we would need a
better classifier to locate an differentiate if from all the other
tokens.